Modify example code for API about Taiwan stock market data or website example.com. [skip ci]
